PythonProject/requirements.txt
azure-mgmt-resource
azure-mgmt-compute
azure-mgmt-network
azure-identity
azure-mgmt-resource~=2.0
azure-mgmt-network~=2.2
azure-mgmt-compute~=4.3
haikunator
azure.common.client_factory
